requirements.txt
pillow
ttkbootstrap
tkinter
requests